feat: add new json files
