+++ crates/core/src/mock/fixtures/requests/blocks/pending/starknet_getBlockWithTxs_pending.json
@@ -1,0 +1,10 @@
+{
+  "id": 1,
+  "jsonrpc": "2.0",
+  "method": "starknet_getBlockWithTxHashes",
+  "params": {
+    "block_id": {
+      "block_hash": "0x0197be2810df6b5eedd5d9e468b200d0b845b642b81a44755e19047f08cc8c6e"
+    }
+  }
+}
