+++ crates/core/src/mock/fixtures/requests/blocks/starknet_getBlockWithTxHashes.json
@@ -1,0 +1,10 @@
+{
+  "id": 1,
+  "jsonrpc": "2.0",
+  "method": "starknet_getBlockWithTxHashes",
+  "params": {
+    "block_id": {
+      "block_tag": "latest"
+    }
+  }
+}
