+++ crates/core/src/mock/fixtures/requests/blocks/starknet_getBlockWithTxs.json
@@ -1,0 +1,10 @@
+{
+  "id": 1,
+  "jsonrpc": "2.0",
+  "method": "starknet_getBlockWithTxs",
+  "params": {
+    "block_id": {
+      "block_tag": "latest"
+    }
+  }
+}
