+++ crates/core/src/mock/fixtures/requests/transactions/starknet_getTransactionByBlockIdAndIndex.json
@@ -1,0 +1,13 @@
+{
+  "id": 1,
+  "jsonrpc": "2.0",
+  "method": "starknet_getTransactionByBlockIdAndIndex",
+  "params": [
+    {
+      "block_id": {
+        "block_tag": "latest"
+      }
+    },
+    0
+  ]
+}
