+++ crates/core/src/mock/mod.rs
@@ -2,4 +2,3 @@ pub mod assert_helpers;
 pub mod constants;
 pub mod mock_starknet;
 pub mod serde;
-pub mod wiremock_utils;
