+++ crates/core/src/models/block.rs
@@ -277,11 +277,10 @@ impl ConvertibleStarknetBlock for BlockWithTxs {
 mod tests {
 
     use super::*;
-    use crate::client::tests::init_mock_client;
     use crate::mock::constants::{
         ABDEL_STARKNET_ADDRESS_HEX, OTHER_ADDRESS_HEX, OTHER_PROXY_ACCOUNT_CLASS_HASH_HEX, PROXY_ACCOUNT_CLASS_HASH_HEX,
     };
-    use crate::mock::mock_starknet::{fixtures, AvailableFixtures};
+    use crate::mock::mock_starknet::{fixtures, init_mock_client, AvailableFixtures};
 
     #[tokio::test]
     async fn test_to_eth_block_block_with_tx_hashes() {
