+++ crates/eth-rpc/tests/utils.rs
@@ -1,13 +1,9 @@
 use std::sync::Arc;
 
+use dojo_test_utils::rpc::MockJsonRpcTransport;
 use kakarot_rpc::servers::eth_rpc::KakarotEthRpc;
-use kakarot_rpc_core::client::config::{JsonRpcClientBuilder, Network, StarknetConfig};
-use kakarot_rpc_core::client::KakarotClient;
-use kakarot_rpc_core::mock::wiremock_utils::setup_wiremock;
-use starknet::core::types::FieldElement;
-use starknet::providers::jsonrpc::HttpTransport;
+use kakarot_rpc_core::mock::mock_starknet::{all_fixtures, init_mock_client};
 use starknet::providers::JsonRpcClient;
-use url::Url;
 
 /// Run wiremock to fake starknet rpc and then run our own `kakarot_rpc_server`.
 ///
@@ -34,21 +30,8 @@ use url::Url;
 ///        let _has_stop = server_handle.stop().unwrap();
 ///   }
 /// ```
-pub async fn setup_kakarot_eth_rpc() -> KakarotEthRpc<JsonRpcClient<HttpTransport>> {
-    let provider_url = setup_wiremock().await;
-    let kakarot_address =
-        FieldElement::from_hex_be("0x566864dbc2ae76c2d12a8a5a334913d0806f85b7a4dccea87467c3ba3616e75").unwrap();
-    let proxy_account_class_hash =
-        FieldElement::from_hex_be("0x0775033b738dfe34c48f43a839c3d882ebe521befb3447240f2d218f14816ef5").unwrap();
-
-    let config = StarknetConfig::new(
-        Network::JsonRpcProvider(Url::parse(&provider_url).unwrap()),
-        kakarot_address,
-        proxy_account_class_hash,
-    );
-    let starknet_provider = JsonRpcClientBuilder::with_http(&config).unwrap().build();
-
-    let kakarot_client = KakarotClient::new(config, starknet_provider);
+pub async fn setup_mock_eth_rpc() -> KakarotEthRpc<JsonRpcClient<MockJsonRpcTransport>> {
+    let kakarot_client = init_mock_client(Some(all_fixtures()));
 
     KakarotEthRpc::new(Arc::new(kakarot_client))
 }
